Mock adapter interface
