+++ .gitignore
@@ -1,2 +1,2 @@
-/target
-/.idea
+target
+.idea
