--- com-api-example/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "com-api-example"
-version = "0.1.0"
-edition = "2024"
-publish = ["common"]
-license = "Apache-2.0"
-
-[dependencies]
-
-[dev-dependencies]
-com-api = { path = "../com-api" }
-com-api-sample-gen = { path = "../com-api-sample-gen" }
-com-api-sample-runtime = { path = "../com-api-sample-runtime" }
-tokio = { version = "~1.43", features = ["rt-multi-thread", "macros", "test-util"] }
