--- com-api-sample-runtime/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "com-api-sample-runtime"
-version = "0.1.0"
-edition = "2024"
-publish = ["common"]
-
-[dependencies]
-com-api = { path = "../com-api" }
-
-[dev-dependencies]
-futures = "0.3"
