+++ com-api/examples/com-api-gen/com-api-gen-mock/src/lib.rs
@@ -22,12 +22,8 @@
 //!
 //! ```
 
-use com_api::{
-    Builder, Consumer, ConsumerBuilder, Interface, OfferedProducer, Producer, ProducerBuilder,
-    Reloc,
-};
-use com_api_sample_runtime::RuntimeImpl;
-use com_api_sample_runtime::{SampleConsumerBuilder, SampleProducerBuilder};
+use com_api::*;
+use com_api_runtime_mock::{RuntimeImpl, SampleConsumerBuilder, SampleProducerBuilder};
 
 #[derive(Debug)]
 pub struct Tire {}
@@ -57,8 +53,8 @@ impl Producer for VehicleProducer {
 }
 
 pub struct VehicleOfferedProducer {
-    pub left_tire: com_api_sample_runtime::Publisher<Tire>,
-    pub exhaust: com_api_sample_runtime::Publisher<Exhaust>,
+    pub left_tire: com_api_runtime_mock::Publisher<Tire>,
+    pub exhaust: com_api_runtime_mock::Publisher<Exhaust>,
 }
 
 impl OfferedProducer for VehicleOfferedProducer {
@@ -82,8 +78,8 @@ impl ProducerBuilder<VehicleInterface, RuntimeImpl, VehicleProducer>
 }
 
 pub struct VehicleConsumer {
-    pub left_tire: com_api_sample_runtime::SubscribableImpl<Tire>,
-    pub exhaust: com_api_sample_runtime::SubscribableImpl<Exhaust>,
+    pub left_tire: com_api_runtime_mock::SubscribableImpl<Tire>,
+    pub exhaust: com_api_runtime_mock::SubscribableImpl<Exhaust>,
 }
 
 impl Consumer for VehicleConsumer {}
